Backspace fix for broken terminals
Back-space key often did not work corretly. This was because
'backspace' key code in the terminfo database was set to something
else then the key actually sends. At the same time 'erase' character
is properly set in the terminal settings, so the key works in other
application.

This patch implements a workaround: if there is no key binding for
a pressed key and the key is current terminal 'erase' character, treat
it as KEY_BACKSPACE. This should give the expected behaviour in most
cases.

As the 'erase' character, sent by the backspace key, is usually '^?' I
had also removed the default binding for '^?'. I hope this won't break
more than this patch is supposed to fix.